# m-automaticontratos-api/swagger_server/models/request_create_suspendidos.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.cliente_suspendido import ClienteSuspendido  # noqa: F401,E501
from swagger_server import util


class RequestCreateSuspendidos(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, external_transaction_id: str=None, channel: str=None, suspendido: ClienteSuspendido=None):  # noqa: E501
        """RequestCreateSuspendidos - a model defined in Swagger

        :param external_transaction_id: The external_transaction_id of this RequestCreateSuspendidos.  # noqa: E501
        :type external_transaction_id: str
        :param channel: The channel of this RequestCreateSuspendidos.  # noqa: E501
        :type channel: str
        :param suspendido: The suspendido of this RequestCreateSuspendidos.  # noqa: E501
        :type suspendido: ClienteSuspendido
        """
        self.swagger_types = {
            'external_transaction_id': str,
            'channel': str,
            'suspendido': ClienteSuspendido
        }

        self.attribute_map = {
            'external_transaction_id': 'externalTransactionId',
            'channel': 'channel',
            'suspendido': 'suspendido'
        }
        self._external_transaction_id = external_transaction_id
        self._channel = channel
        self._suspendido = suspendido

    @classmethod
    def from_dict(cls, dikt) -> 'RequestCreateSuspendidos':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The RequestCreateSuspendidos of this RequestCreateSuspendidos.  # noqa: E501
        :rtype: RequestCreateSuspendidos
        """
        return util.deserialize_model(dikt, cls)

    @property
    def external_transaction_id(self) -> str:
        """Gets the external_transaction_id of this RequestCreateSuspendidos.


        :return: The external_transaction_id of this RequestCreateSuspendidos.
        :rtype: str
        """
        return self._external_transaction_id

    @external_transaction_id.setter
    def external_transaction_id(self, external_transaction_id: str):
        """Sets the external_transaction_id of this RequestCreateSuspendidos.


        :param external_transaction_id: The external_transaction_id of this RequestCreateSuspendidos.
        :type external_transaction_id: str
        """
        if external_transaction_id is None:
            raise ValueError("Invalid value for `external_transaction_id`, must not be `None`")  # noqa: E501

        self._external_transaction_id = external_transaction_id

    @property
    def channel(self) -> str:
        """Gets the channel of this RequestCreateSuspendidos.


        :return: The channel of this RequestCreateSuspendidos.
        :rtype: str
        """
        return self._channel

    @channel.setter
    def channel(self, channel: str):
        """Sets the channel of this RequestCreateSuspendidos.


        :param channel: The channel of this RequestCreateSuspendidos.
        :type channel: str
        """
        if channel is None:
            raise ValueError("Invalid value for `channel`, must not be `None`")  # noqa: E501

        self._channel = channel

    @property
    def suspendido(self) -> ClienteSuspendido:
        """Gets the suspendido of this RequestCreateSuspendidos.


        :return: The suspendido of this RequestCreateSuspendidos.
        :rtype: ClienteSuspendido
        """
        return self._suspendido

    @suspendido.setter
    def suspendido(self, suspendido: ClienteSuspendido):
        """Sets the suspendido of this RequestCreateSuspendidos.


        :param suspendido: The suspendido of this RequestCreateSuspendidos.
        :type suspendido: ClienteSuspendido
        """
        if suspendido is None:
            raise ValueError("Invalid value for `suspendido`, must not be `None`")  # noqa: E501

        self._suspendido = suspendido